Remove tundrastore constants.py, relocate constants to their users
Moves DoorState/NestState into standard.py (next to the DoorStatus/
NestStatus dataclasses that use them) and the protocol reply tokens
(ACK!/OK!/ABORTED!/ERROR!) into backend.py, their only consumer. Drops
the unused PLATE_AVAILABLE constant and deletes the now-empty module.
