+++ pylabrobot/highres/tundrastore/__init__.py
@@ -1,6 +1,5 @@
 from .backend import TundraStoreBackend
 from .chatterbox import TundraStoreChatterboxBackend
-from .constants import DoorState, NestState
 from .errors import (
   PlateNotFoundError,
   TundraStoreAbortedError,
@@ -9,8 +8,10 @@ from .errors import (
 )
 from .settings import TundraStoreSettings
 from .standard import (
+  DoorState,
   DoorStatus,
   EnvironmentParameter,
+  NestState,
   NestStatus,
   StackerDimensions,
   VersionInfo,
